Circumvents the problem of #1
This commit aims to circumvent the problem noted in issue #1 (for commit 333fffff93d9c5ff263c6bc83ffeec47d520a76f) by removing the need for finding an initial point within the polygon. It chooses the origin (which is assured to stay within the polygon). Since all starting points yield the same end-result, this solution seems quite feasible and justified, as it also saves a lot of computational resources.